requirements.txt
eventlet
loguru
python-socketio